parses out url and query string information from request.
