+++ http-start.js
@@ -1,0 +1,2 @@
+const app = require('./app');
+app.listen(4444);
