+++ index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>Chatty Cow</title>
+</head>
+<body>
+
+  <h1>Talk to Chatty Cow</h1>
+
+	<form method='POST' >
+		<input name="name" value="talk to me">
+		<button type="submit">send</button>
+	</form>
+
+</body>
+</html>
